split state.ts off of reducer.ts
